pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "heart-murmur-analysis"
version = "0.1.6"
description = "Heartbeat analysis using Deep Learning, Signal Processing, and AI Agents."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Aanand", email = "aanan@example.com"}
]
dependencies = [
    "streamlit",
    "python-docx",
    "numpy<2.0",
    "pandas",
    "matplotlib",
    "scikit-learn",
    "tensorflow",
    "tf-keras",
    "librosa==0.10.2",
    "numba>=0.59.0",
    "soundfile",
    "python-dotenv",
    "pydantic",
    "langchain-groq",
    "langchain",
    "langchain-community",
    "langchain_google_genai",
    "langchain-text-splitters",
    "chromadb",
    "faiss-cpu",
    "openpyxl",
    "langchain-core",
    "langchain-huggingface",
    "langgraph",
    "sentence-transformers"
]

[project.scripts]
heart-murmur-app = "heart_murmur_analysis.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["heart_murmur_analysis"]

[tool.hatch.build.targets.sdist]
include = [
    "heart_murmur_analysis",
    "README.md",
]

[tool.hatch.build]
artifacts = [
    "heart_murmur_analysis/models/*.h5"
]
